Switch to extending from OsmApi.
Waiting for https://josm.openstreetmap.de/ticket/9744 to get merged in.
