+++ src/org/openstreetmap/josm/plugins/notes/api/CloseAction.java
@@ -30,12 +30,12 @@ package org.openstreetmap.josm.plugins.notes.api;
 import org.openstreetmap.josm.Main;
 import org.openstreetmap.josm.io.OsmTransferException;
 import org.openstreetmap.josm.plugins.notes.Note;
-import org.openstreetmap.josm.plugins.notes.api.util.NotesApi;
+import org.openstreetmap.josm.plugins.notes.api.util.NotesCapableOsmApi;
 
 public class CloseAction {
 
     public void execute(Note n, String closeMsg) throws OsmTransferException {
-		NotesApi.getNotesApi().closeNote(n, closeMsg);
+		NotesCapableOsmApi.getNotesApi().closeNote(n, closeMsg);
 		n.setState(Note.State.closed);
 		Main.map.mapView.repaint();
     }
