+++ src/org/openstreetmap/josm/plugins/notes/gui/action/ReopenAction.java
@@ -11,7 +11,7 @@ import org.openstreetmap.josm.gui.widgets.HistoryChangedListener;
 import org.openstreetmap.josm.plugins.notes.ConfigKeys;
 import org.openstreetmap.josm.plugins.notes.Note;
 import org.openstreetmap.josm.plugins.notes.NotesPlugin;
-import org.openstreetmap.josm.plugins.notes.api.util.NotesApi;
+import org.openstreetmap.josm.plugins.notes.api.util.NotesCapableOsmApi;
 import org.openstreetmap.josm.plugins.notes.gui.NotesDialog;
 import org.openstreetmap.josm.plugins.notes.gui.dialogs.TextInputDialog;
 
@@ -54,7 +54,7 @@ public class ReopenAction extends NotesAction {
 		if(note == null) {
 			return;
 		}
-		NotesApi.getNotesApi().reopenNote(note, comment);
+        NotesCapableOsmApi.getNotesApi().reopenNote(note, comment);
 		note.setState(Note.State.open);
 		Main.map.mapView.repaint();
 	}
